ハッシュ使用率のオーバーフロー修正

+++ usi/ZobristHash.h
@@ -46,7 +46,7 @@ public:
 	bool CheckRemainingHashSize(void) const { return enough_size; }
 
 	// ハッシュ使用率を取得(単位はパーミル(全体を1000とした値))
-	int GetUctHashUsageRate() const { return 1000 * used / uct_hash_size; }
+	int GetUctHashUsageRate() const { return (int)(1000.0 * used / uct_hash_size); }
 
 	// ノードを返す
 	const node_hash_t& operator [](const size_t i) { return node_hash[i]; }
